@@ -1,22 +1,8 @@
-import React, { Fragment, useState, useEffect } from "react";
-
-import "react-toastify/dist/ReactToastify.css";
-import {
-  BrowserRouter as Router,
-  Route,
-  Switch,
-  Redirect
-} from "react-router-dom";
-
-import { toast } from "react-toastify";
-
+import React, { useState, useEffect } from "react";
+import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom'
+import './App.css';
 //components
-
-import Login from "./components/Login";
-import Register from "./components/Register";
-import Dashboard from "./components/Dashboard";
-
-toast.configure();
+import { Dashboard, Login, Register } from "./components";
 
 function App() {
   const checkAuthenticated = async () => {
@@ -43,49 +29,19 @@ function App() {
   const setAuth = boolean => {
     setIsAuthenticated(boolean);
   };
-
   return (
-    <Fragment>
-      <Router>
-        <div className="container">
-          <Switch>
-            <Route
-              exact
-              path="/login"
-              render={props =>
-                !isAuthenticated ? (
-                  <Login {...props} setAuth={setAuth} />
-                ) : (
-                  <Redirect to="/dashboard" />
-                )
-              }
-            />
-            <Route
-              exact
-              path="/register"
-              render={props =>
-                !isAuthenticated ? (
-                  <Register {...props} setAuth={setAuth} />
-                ) : (
-                  <Redirect to="/dashboard" />
-                )
-              }
-            />
-            <Route
-              exact
-              path="/dashboard"
-              render={props =>
-                isAuthenticated ? (
-                  <Dashboard {...props} setAuth={setAuth} />
-                ) : (
-                  <Redirect to="/login" />
-                )
-              }
-            />
-          </Switch>
-        </div>
-      </Router>
-    </Fragment>
+    <Router>
+      <div className="container">
+        <Routes>
+          <Route exact path="/login" element={!isAuthenticated ? (
+            <Login setAuth={setAuth} />) : (<Navigate to="/dashboard" />)} />
+          <Route exact path="/register" element={!isAuthenticated ? (<Register setAuth={setAuth} />) :
+            (<Navigate to="/dashboard" />)} />
+          <Route exact path="/dashboard" element={isAuthenticated ? (
+            <Dashboard setAuth={setAuth} />) : (<Navigate to="/login" />)} />
+        </Routes>
+      </div>
+    </Router>
   );
 }
 
