@@ -1,6 +1,6 @@
-import React, { Fragment, useState } from "react";
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
-
+import "../App.css";
 import { toast } from "react-toastify";
 
 const Login = ({ setAuth }) => {
@@ -44,8 +44,9 @@ const Login = ({ setAuth }) => {
     }
   };
 
+
   return (
-    <Fragment>
+    <main>
       <h1 className="mt-5 text-center">Login</h1>
       <form onSubmit={onSubmitForm}>
         <input
@@ -64,8 +65,8 @@ const Login = ({ setAuth }) => {
         />
         <button className="btn btn-success btn-block">Submit</button>
       </form>
-      <Link to="/register">register</Link>
-    </Fragment>
+      <Link to="/register" className="btn btn-warning">register</Link>
+    </main>
   );
 };
 
