@@ -1,7 +1,7 @@
 const jwt = require("jsonwebtoken");
 require("dotenv").config();
 
-//this middleware will only continue if the token is inside the local storage
+//this middleware will on continue on if the token is inside the local storage
 
 module.exports = function(req, res, next) {
   // Get token from header
